Split up bokeh/models/annotations.py
