+++ bokeh/models/__init__.py
@@ -32,8 +32,7 @@ log = logging.getLogger(__name__)
 # Bokeh imports
 from ..core.property.dataspec import expr, field, value  # Legacy API
 from ..model import Model
-from .annotations import *
-from .arrow_heads import *
+from .annotations import *  # lgtm [py/polluting-import]
 from .axes import *
 from .callbacks import *
 from .canvas import *
