--- sphinx/source/docs/reference/models/arrow_heads.rst
@@ -1,8 +1,0 @@
-.. _bokeh.models.arrow_heads:
-
-arrow_heads
------------
-
-.. automodule:: bokeh.models.arrow_heads
-   :members:
-   :undoc-members:
